requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
rank-bm25>=0.2.2
nltk>=3.8
numpy>=1.21.0
requests>=2.28.0